+++ lib/active_record/connection_adapters/sqlserver/database_statements.rb
@@ -149,7 +149,12 @@ module ActiveRecord
           sql = +"INSERT #{insert.into}"
 
           if returning = insert.send(:insert_all).returning
-            sql << " OUTPUT " << returning.map { |column| "INSERTED.#{quote_column_name(column)}" }.join(", ")
+            returning_sql = if returning.is_a?(String)
+              returning
+            else
+              returning.map { |column| "INSERTED.#{quote_column_name(column)}" }.join(", ")
+            end
+            sql << " OUTPUT #{returning_sql}"
           end
 
           sql << " #{insert.values_list}"